# specs/002-meeting-room-booking/system-analyze/_prompt-consistency/openapi.run-C.yaml
openapi: 3.0.3
info:
  title: 內部共享會議室預約系統 API
  version: 0.1.0-draft
  description: |
    功能分支：`002-meeting-room-booking`

    **狀態**：草稿

    ## API Schema 描述

    | 欄位 | 內容 |
    | --- | --- |
    | 共通 Header | JSON request 使用 `Content-Type: application/json`；登入後由瀏覽器自動帶入 HttpOnly session cookie；所有受保護 Endpoint 以該 session 辨識使用者 |
    | 時間格式 | 瞬間使用 ISO-8601 with offset（例如 `2026-07-31T14:00:00+08:00`）或 UTC `Z`；日期 query 使用 `YYYY-MM-DD`，所有「今日／平日／營業時段」以 `Asia/Taipei` 解讀 |

    ## 假設

    - 所有 `/api` Endpoint 除 `POST /api/auth/login` 外皆要求有效 session cookie；未登入固定回 `401 UNAUTHENTICATED`
    - 管理操作只允許 `facility_admin`；`employee` 與 `manager` 權限相同，越權固定回 `403 FORBIDDEN`
    - 時段採半開區間 `[starts_at, ends_at)`；相鄰但不重疊的預約或維護時段可成立
    - `POST /api/bookings` 的停用、容量、跨日、週末／假日、營業時段與時長違規使用 `400 BOOKING_RULE_VIOLATION`；設備需求欄位僅記錄；預約或維護時段重疊使用 409
    - 日曆與列表直接共用 `GET /api/bookings` 回應；第一版不另設 view 參數
    - `GET /api/bookings/mine` 第一版回傳全部歷史，不做分頁；預設依開始時間降冪排序
    - `GET /api/overview/today` 的 `busy_ratio` 以平日營業窗 09:00–21:00 共 720 分鐘計算；週末或假日的 `business_minutes` 與 `busy_ratio` 皆為 0
    - 維護時段查詢對所有已登入角色開放，讓預約畫面可先提示不可用；建立維護時段仍限設施管理員
...
paths:
...
  /api/rooms:
    get:
      tags:
        - Rooms
      summary: 列出所有會議室
      description: |
        列出所有會議室，包含啟用與停用項目。

        **對應 FR**：US1-FR2（顯示會議室名稱、樓層、容量、設備及啟用狀態）

        **設計備註**：不預設排除停用會議室；前端以 `is_active` 明確呈現不可新約狀態；預設依 `floor`、`name` 排序。

        **測試規劃**
        - 清單同時含啟用與停用會議室 → 200
        - 尚無會議室時回空陣列 → 200
        - 未登入查詢 → 401
      operationId: listRooms
      responses:
        '200':
          description: 成功回傳會議室清單
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomListResponse'
              example:
                rooms:
                  - id: '22222222-2222-4222-8222-222222222222'
                    name: 大會議A
                    floor: 3F
                    capacity: 12
                    has_projector: true
                    has_video_conference: true
                    is_active: true
                    created_at: '2026-07-30T01:00:00.000Z'
                    updated_at: '2026-07-30T01:00:00.000Z'
                  - id: '66666666-6666-4666-8666-666666666666'
                    name: 小會議B
                    floor: 4F
                    capacity: 6
                    has_projector: false
                    has_video_conference: true
                    is_active: false
                    created_at: '2026-07-30T01:10:00.000Z'
                    updated_at: '2026-07-30T04:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthenticated'

    post:
      tags:
        - Rooms
      summary: 新增會議室
      description: |
        建立一間預設啟用的會議室。

        **對應 FR**
        - US4-FR1：設施管理員可新增會議室
        - US4-FR4：拒絕非設施管理員執行管理操作

        **設計備註**：僅 `facility_admin`；`name`、`floor` 不可空白，`capacity` 至少為 1；`has_projector` 與 `has_video_conference` 必填且須為布林值；建立時 `is_active` 固定為 `true`。

        **測試規劃**
        - 設施管理員建立會議室 → 201
        - 欄位缺漏或容量無效 → 400
        - 未登入建立 → 401
        - 員工或主管建立 → 403
      operationId: createRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
            example:
              name: 研討室C
              floor: 5F
              capacity: 8
              has_projector: true
              has_video_conference: false
      responses:
        '201':
          description: 會議室建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
              example:
                id: '77777777-7777-4777-8777-777777777777'
                name: 研討室C
                floor: 5F
                capacity: 8
                has_projector: true
                has_video_conference: false
                is_active: true
                created_at: '2026-07-30T05:00:00.000Z'
                updated_at: '2026-07-30T05:00:00.000Z'
        '400':
          description: 欄位缺漏或容量無效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: VALIDATION_ERROR
                  message: capacity must be at least 1
                  details:
                    - field: capacity
                      reason: must be at least 1
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/FacilityAdminRequired'
  /api/rooms/{roomId}:
    patch:
      tags:
        - Rooms
      summary: 停用會議室
      description: |
        更新指定會議室的啟用狀態。

        **對應 FR**
        - US4-FR2：設施管理員可將會議室停用
        - US4-FR4：拒絕非設施管理員執行停用操作
        - GR-003：停用只禁止新預約，不取消既有已確認預約

        **設計備註**：僅 `facility_admin`；第一版 request 僅接受 `is_active: false`，不開放重新啟用；停用只擋新預約，不取消或改動既有預約。

        **測試規劃**
        - 設施管理員停用會議室 → 200
        - is_active 缺漏、格式錯誤或傳入 true → 400
        - 未登入更新 → 401
        - 員工或主管更新 → 403
        - roomId 不存在 → 404
      operationId: deactivateRoom
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeactivateRoomRequest'
            example:
              is_active: false
      responses:
        '200':
          description: 會議室停用成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
              example:
                id: '22222222-2222-4222-8222-222222222222'
                name: 大會議A
                floor: 3F
                capacity: 12
                has_projector: true
                has_video_conference: true
                is_active: false
                created_at: '2026-07-30T01:00:00.000Z'
                updated_at: '2026-07-30T05:30:00.000Z'
        '400':
          description: is_active 缺漏、格式錯誤或傳入 true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: VALIDATION_ERROR
                  message: is_active must be false
                  details:
                    - field: is_active
                      reason: only deactivation is supported
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/FacilityAdminRequired'
        '404':
          $ref: '#/components/responses/RoomNotFound'
...
  /api/rooms/{roomId}/maintenance-windows:
    post:
      tags:
        - Maintenance
      summary: 新增維護時段
      description: |
        為指定會議室新增不可建立預約的維護區間。

        **對應 FR**
        - US4-FR3：設施管理員可設定會議室維護黑名單時段
        - US4-FR4：拒絕非設施管理員設定維護時段

        **設計備註**：僅 `facility_admin`；`created_by` 取自 session；`ends_at` 必須晚於 `starts_at`；維護區間採半開區間 `[starts_at, ends_at)`；不自動取消既有 confirmed 預約。

        **測試規劃**
        - 設施管理員新增有效維護時段 → 201
        - 時間格式錯誤或結束不晚於開始 → 400
        - 未登入新增 → 401
        - 員工或主管新增 → 403
        - roomId 不存在 → 404
      operationId: createMaintenanceWindow
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaintenanceWindowRequest'
            example:
              starts_at: '2026-08-01T09:00:00+08:00'
              ends_at: '2026-08-01T12:00:00+08:00'
              note: 空調維修
      responses:
        '201':
          description: 維護時段建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceWindow'
              example:
                id: '44444444-4444-4444-8444-444444444444'
                room_id: '22222222-2222-4222-8222-222222222222'
                starts_at: '2026-08-01T09:00:00+08:00'
                ends_at: '2026-08-01T12:00:00+08:00'
                note: 空調維修
                created_by: '55555555-5555-4555-8555-555555555555'
                created_at: '2026-07-30T03:00:00.000Z'
        '400':
          description: 時間格式錯誤或結束不晚於開始
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: VALIDATION_ERROR
                  message: ends_at must be later than starts_at
                  details:
                    - field: ends_at
                      reason: must_be_after_starts_at
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/FacilityAdminRequired'
        '404':
          $ref: '#/components/responses/RoomNotFound'

    get:
      tags:
        - Maintenance
      summary: 查詢維護時段
      description: |
        查詢指定會議室在時間區間內相交的維護時段。

        **對應 FR**
        - US1-FR5：支援預約前辨識維護不可用時段
        - US4-FR3：支援管理員查閱指定會議室維護時段

        **設計備註**：已登入三角色皆可查詢，支援預約 UI 預檢與管理畫面；`from`、`to` 皆必填且 `to` 晚於 `from`；結果依 `starts_at` 排序。

        **測試規劃**
        - 查詢區間內有維護時段 → 200
        - 查詢區間內無維護時段時回空陣列 → 200
        - from／to 缺漏、格式錯誤或順序無效 → 400
        - 未登入查詢 → 401
        - roomId 不存在 → 404
      operationId: listMaintenanceWindows
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
        - name: from
          in: query
          required: true
          description: 查詢區間起始瞬間（ISO-8601 with offset 或 UTC Z）
          schema:
            type: string
            format: date-time
          example: '2026-08-01T00:00:00+08:00'
        - name: to
          in: query
          required: true
          description: 查詢區間結束瞬間，必須晚於 `from`
          schema:
            type: string
            format: date-time
          example: '2026-08-02T00:00:00+08:00'
      responses:
        '200':
          description: 成功回傳維護時段清單
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceWindowListResponse'
              example:
                room_id: '22222222-2222-4222-8222-222222222222'
                maintenance_windows:
                  - id: '44444444-4444-4444-8444-444444444444'
                    room_id: '22222222-2222-4222-8222-222222222222'
                    starts_at: '2026-08-01T09:00:00+08:00'
                    ends_at: '2026-08-01T12:00:00+08:00'
                    note: 空調維修
                    created_by: '55555555-5555-4555-8555-555555555555'
                    created_at: '2026-07-30T03:00:00.000Z'
        '400':
          description: from／to 缺漏、格式錯誤或順序無效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: VALIDATION_ERROR
                  message: from and to must be valid ISO-8601 datetimes and to must be later than from
                  details: []
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/RoomNotFound'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: |
        HttpOnly session cookie，由 `connect-pg-simple` 管理。
        登入成功後瀏覽器自動帶入；所有受保護 Endpoint 以該 session 辨識使用者。
        僅 `POST /api/auth/login` 為公開端點，不需此 cookie。
...
  parameters:
    RoomIdPath:
      name: roomId
      in: path
      required: true
      description: 會議室 UUID
      schema:
        $ref: '#/components/schemas/Uuid'
      example: '22222222-2222-4222-8222-222222222222'
...
  responses:
    Unauthenticated:
      description: 未登入或 session 無效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHENTICATED
              message: Authentication is required
              details: []
    FacilityAdminRequired:
      description: 需要設施管理員角色
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: Facility administrator role is required
              details: []
    RoomNotFound:
      description: 會議室不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: ROOM_NOT_FOUND
              message: Room was not found
              details: []
...
  schemas:
    Uuid:
      type: string
      format: uuid
      description: 對外資源 ID（UUID 字串）
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      description: |
        全檔共用的錯誤回應格式。
        各 Endpoint 仍逐一列出可能的 error status 與完整錯誤範例。
        `details` 可用於提供欄位或規則細節。
    ErrorBody:
      type: object
      required:
        - code
        - message
        - details
      properties:
        code:
          type: string
          description: 機器可讀的錯誤代碼
          enum:
            - VALIDATION_ERROR
            - INVALID_CREDENTIALS
            - UNAUTHENTICATED
            - FORBIDDEN
            - ROOM_NOT_FOUND
            - BOOKING_NOT_FOUND
            - BOOKING_RULE_VIOLATION
            - BOOKING_CONFLICT
            - MAINTENANCE_CONFLICT
            - BOOKING_NOT_CANCELLABLE
        message:
          type: string
          description: 人類可讀的錯誤訊息
        details:
          type: array
          description: 欄位、規則或衝突細節
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      description: |
        錯誤細節項目，依情境包含不同欄位：
        - 驗證錯誤：`field` + `reason`
        - 預約規則違反：`rule` + `reason`
        - 時段衝突：`room_id` + `starts_at` + `ends_at`
      oneOf:
        - $ref: '#/components/schemas/FieldErrorDetail'
        - $ref: '#/components/schemas/RuleErrorDetail'
        - $ref: '#/components/schemas/ConflictErrorDetail'
    FieldErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: 驗證失敗的欄位名稱
        reason:
          type: string
          description: 驗證失敗原因
      required:
        - field
        - reason
      additionalProperties: false
    RuleErrorDetail:
      type: object
      properties:
        rule:
          type: string
          description: 違反的規則代碼（例如 GR-002）
        reason:
          type: string
          description: 規則違反原因
      required:
        - rule
        - reason
      additionalProperties: false
    ConflictErrorDetail:
      type: object
      properties:
        room_id:
          $ref: '#/components/schemas/Uuid'
        starts_at:
          type: string
          format: date-time
          description: 衝突時段開始瞬間
        ends_at:
          type: string
          format: date-time
          description: 衝突時段結束瞬間
      required:
        - room_id
        - starts_at
        - ends_at
      additionalProperties: false
...
    Room:
      type: object
      required:
        - id
        - name
        - floor
        - capacity
        - has_projector
        - has_video_conference
        - is_active
        - created_at
        - updated_at
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
          description: 會議室 UUID
        name:
          type: string
          description: 會議室名稱
        floor:
          type: string
          description: 樓層標示
        capacity:
          type: integer
          minimum: 1
          description: 最大容納人數
        has_projector:
          type: boolean
          description: 是否具備投影機
        has_video_conference:
          type: boolean
          description: 是否具備視訊設備
        is_active:
          type: boolean
          description: 是否接受新預約；停用會議室仍保留並可辨識
        created_at:
          type: string
          format: date-time
          description: 建立時間
        updated_at:
          type: string
          format: date-time
          description: 最後更新時間
    RoomListResponse:
      type: object
      required:
        - rooms
      properties:
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room'
    CreateRoomRequest:
      type: object
      required:
        - name
        - floor
        - capacity
        - has_projector
        - has_video_conference
      properties:
        name:
          type: string
          minLength: 1
          description: 會議室名稱，不可空白
        floor:
          type: string
          minLength: 1
          description: 樓層標示，不可空白
        capacity:
          type: integer
          minimum: 1
          description: 最大容納人數，至少為 1
        has_projector:
          type: boolean
          description: 是否具備投影機（必填）
        has_video_conference:
          type: boolean
          description: 是否具備視訊設備（必填）
    DeactivateRoomRequest:
      type: object
      required:
        - is_active
      properties:
        is_active:
          type: boolean
          enum:
            - false
          description: 第一版僅接受 `false`（停用），不開放重新啟用
...
    MaintenanceWindow:
      type: object
      required:
        - id
        - room_id
        - starts_at
        - ends_at
        - created_by
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
          description: 維護時段 UUID
        room_id:
          $ref: '#/components/schemas/Uuid'
          description: 所屬會議室 UUID
        starts_at:
          type: string
          format: date-time
          description: 維護開始瞬間
        ends_at:
          type: string
          format: date-time
          description: 維護結束瞬間
        note:
          type: string
          description: 可選的維護說明
        created_by:
          $ref: '#/components/schemas/Uuid'
          description: 建立此維護時段的設施管理員 UUID
        created_at:
          type: string
          format: date-time
          description: 建立時間
    CreateMaintenanceWindowRequest:
      type: object
      required:
        - starts_at
        - ends_at
      properties:
        starts_at:
          type: string
          format: date-time
          description: 維護開始瞬間
        ends_at:
          type: string
          format: date-time
          description: 維護結束瞬間，必須晚於 starts_at
        note:
          type: string
          description: 可選的維護說明
    MaintenanceWindowListResponse:
      type: object
      required:
        - room_id
        - maintenance_windows
      properties:
        room_id:
          $ref: '#/components/schemas/Uuid'
        maintenance_windows:
          type: array
          items:
            $ref: '#/components/schemas/MaintenanceWindow'